pyproject.toml
[project]
name = "ecstasy"
version = "18.3.7"
description = "Ecstasy - Equipment Control System To Assist You - программа для управления сетевым оборудованием."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alive-progress>=3.2.0",
    "celery>=5.5.1",
    "certifi>=2025.1.31",
    "coreapi>=2.3.3",
    "cryptography>=44.0.2",
    "django>=5.2",
    "django-celery-beat>=2.8.0",
    "django-cors-headers>=4.7.0",
    "django-dbbackup>=4.2.1",
    "django-filter>=25.1",
    "django-import-export>=4.3.7",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "django-debug-toolbar>=5.1.0",
    "drf-orjson-renderer>=1.7.3",
    "drf-yasg>=1.21.10",
    "flask>=3.1.0",
    "flower>=2.0.1",
    "gevent>=24.11.1",
    "gunicorn>=23.0.0",
    "idna>=3.10",
    "jinja2>=3.1.6",
    "mysqlclient>=2.2.7",
    "openpyxl>=3.1.5",
    "orjson>=3.10.16",
    "paramiko>=3.5.1",
    "pexpect>=4.9.0",
    "pillow>=11.2.1",
    "ping3>=4.0.8",
    "pysftp>=0.2.9",
    "pyvis==0.3.2",
    "pyyaml>=6.0.2",
    "pyzabbix>=1.3.1",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "sqlparse>=0.5.3",
    "tabulate>=0.9.0",
    "textfsm>=1.1.3",
    "urllib3>=2.4.0",
    "xlwt>=1.3.0",
]

[dependency-groups]
dev = [
    "black[d]>=25.1.0",
    "pdoc>=15.0.2",
    "pylint>=3.3.6",
    "pylint-django>=2.6.1",
    "celery-types>=0.23.0",
    "django-filter-stubs>=0.1.3",
    "django-stubs>=5.1.3",
    "djangorestframework-stubs>=3.15.3",
    "drf-yasg-stubs>=0.1.4",
    "mypy>=1.15.0",
    "types-markdown>=3.8.0.20250415",
    "types-pexpect>=4.9.0.20241208",
    "types-psycopg2>=2.9.21.20250318",
    "types-pygments>=2.19.0.20250305",
    "types-pysftp>=0.2.17.20241117",
    "types-tabulate>=0.9.0.20241207",
]

[tool.ruff]
line-length = 120
lint.extend-select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
lint.ignore = [
    # Line too long
    "E501",
    # Functions with a high complexity
    "C901",
    # Use ternary operator instead of `if`-`else`-block
    "SIM108",
    # Module level import not at top of file
    "E402"
]